Torxakis configuration via config file (#198)
Prepared command line config parsing and interpretation to add a configuration
file parsing in the future.
Updated `stack.yaml` to use the integer simple variants of:
- hashable-1.2.6.1
- integer-logarithms-1.0.2
- scientific-0.3.5.1
Updated CI setup script to make symbolic links to `cvc4`.
Removed dependency of the tests in `solve` on Config. Created a `TestSolvers`
that contains the dependencies. The solve tests need to be refactored to avoid
hardcoding solvers names.