Correct an error when the compiler will find match a process with no parameters. In the example below:
```text
PROCDEF noop [ ] (  ) ::=
    STOP
ENDDEF

PROCDEF noop [ ] ( id :: Int ) ::=
        noop [ ] (10)
ENDDEF
```
the compiler would match both definitions of `noop` since the empty list of val
expressions would be allowed. Now we check for the length of the parameters.

Also an error of `Control.Monad.Error` is thrown when no matching process
definition is found, or when multiple process definitions are found.