Updates Haskell LTS version to 18.28 (#977)
* Fixes several build issues in the ui module.

The Network module is removed, and has been replaced with a combination of the Network.Socket module, and functions from TextViaSockets.
Exception handling is removed from the haskeline library. It is replaced with the exceptions library instead.

* Fixes the valexpr module

The monoid typeclass has split into Monoid and Semigroup, making old implementations of Monoid incorrect.
This is fixed by moving the implementation of mappend to the implementation of <> in a new instance of semigroup

* Remove redundant import Data.Monoid statements

* Fixes serverenv and core modules

Mostly removes redundant monad imports, and replaces the Network library with Network.Socket.

* Fixes server module

Replaces the removed Network module with parts from Network.Socket, and parts from TextOverSockets.

Definately needs testing.

* Fixes test files

Add Semigroup instance, and replace some calls to Network module with calls to Network.Socket module.

* fixed minor typo in build file that got stuck in cached results and broke everything

* Updated integration test ghc version to be the same as main project ghc version

* Sets all resolver versions spread out over the project to the same resolver, and adds notes to each file pointing to the need to keep the other files updated if things change again in the future.

* Fixes integration tests building

tests build now, but they do not execute because they expect the torxakis executable to be on the path, wich messes with global instalations.

* Start of fixing benchmark sanity tests

Try to standardize the working directory for all tests to the main git folder. Some tests were running in the sqatt folder, while others were not.
This makes using relative paths to the torxakis executable impossible.

Currently the benchmark sanity tests are still failing nondeterministically.

* Remove parallel test execution

parallel test execution was giving nondeterministic errors.
Errors seemed to not be related to the body of the tests, as replacing all tests with 1 `shouldbe` 10 still crashed.
This only seemed to happen when at least 3 testsets where given as an argument to the testExampleSets function, but more frequently with more testsets being used.

observed error:
$ stack test sqatt:benchmarks-sanity-test
sqatt> test (suite: benchmarks-sanity-test)

sqatt> Test suite benchmarks-sanity-test failed

Error: [S-7282]
       Stack failed to execute the build plan.

       While executing the build plan, Stack encountered the error:

       Error: [S-1995]
       Test suite failure for package sqatt-0.1
           benchmarks-sanity-test:  exited with: ExitFailure (-1073741675)

expected behaviour: something like
$ stack test sqatt:benchmarks-sanity-test
sqatt> test (suite: benchmarks-sanity-test)

Debug set 1
  dummy test. FAILED [1]
Debug set 1
  dummy test. FAILED [2]
Debug set 1
  dummy test. FAILED [3]

Failures:

  src\Sqatt.hs:497:46:
  1) Debug set 1 dummy test.
       expected: 10
        but got: 1

  To rerun use: --match "/Debug set 1/dummy test./"

  src\Sqatt.hs:497:46:
  2) Debug set 1 dummy test.
       expected: 10
        but got: 1

  To rerun use: --match "/Debug set 1/dummy test./"

  src\Sqatt.hs:497:46:
  3) Debug set 1 dummy test.
       expected: 10
        but got: 1

  To rerun use: --match "/Debug set 1/dummy test./"

Randomized with seed 1632928150

Finished in 0.0006 seconds
3 examples, 3 failures

* Remove more parallel test execution, and support for czc4

* Fixes integration tests.

* Fixes benchmarks, and updates testing readme

* Disabled movingarms test in exploremodels testset.

Test removed as it does not terminate, tested in latest develop, v0.9 and v0.6,

* General fixes to become compatible with newer version of ghc.

Mostly removed redundant imports and error cases.
Temporarily adds Unique as extra dependency, because replacing the depriciated package is probably not worth it if the module in which it is used (txs-compiler) is planned to be phased out.

* Fixes unit tests for newer version of ghc

* Fixes integration of changes into system tests.

* Changed dynamic testset to use same pathing convention as other tests

Default is relative path starting in root git directory.
Relative paths starting in sqatt directory need to use explicit insqatt modifier.

* Fixes typo

* Fix linux stack.yaml to be in line with windows stack.yaml

* Removes integersimple ghc variant

All the tests still pass without it, and the original licensing issue seems to have been resolved.

* Updates version of hspec used

Fixes #978

---------

Co-authored-by: Gijs van Cuyck <gijs.vancuyck@ru.nl>